+++ src/sll/include/sll/sandbox.h
@@ -187,6 +187,18 @@
 
 
 /**
+ * \flags macro var
+ * \name SLL_SANDBOX_FLAG_DISABLE_SERIAL
+ * \group sandbox
+ * \subgroup sandbox-flag
+ * \desc Docs!
+ * \type sll_sandbox_flag_t
+ */
+#define SLL_SANDBOX_FLAG_DISABLE_SERIAL 14
+
+
+
+/**
  * \flags check_output func
  * \name sll_get_sandbox_flag
  * \group sandbox